requirements.txt
transformers
openai
pandas
numpy
python-dotenv
pytube
re
streamlit